requirements.txt
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
